requirements.txt
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.9.23
fsspec==2025.10.0
ftfy==6.3.1
grpcio==1.76.0
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marker-pdf==0.2.5
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.19.2
opencv-python==4.11.0.86
packaging==25.0
pdfplumber==0.11.8
pdftext==0.3.14
pillow==10.4.0
pypdf==6.2.0
protobuf==6.33.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pypdfium2==4.30.0
python-dotenv==1.2.1
PyYAML==6.0.3
RapidFuzz==3.13.0
ray==2.49.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
surya-ocr==0.4.5
sympy==1.14.0
tabulate==0.9.0
texify==0.1.10
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.67.1
transformers==4.36.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
zipp==3.23.0
